requirements.txt
django==1.9
djangorestframework==3.3.3
mongoengine==0.9
pymongo==2.7
django-rest-framework-mongoengine
django-cors-headers==2.1.0